backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
anthropic==0.18.1
python-dotenv==1.0.0
python-multipart==0.0.6
pydantic==2.5.3
aiofiles==23.2.1
requests==2.31.0
beautifulsoup4==4.12.3
lxml==5.1.0